@@ -1,5 +1,4 @@
 import { useState } from 'react'
-import { Button } from "@/components/ui/button"
 
 export default function AdoptionList() {
   const [count, setCount] = useState(0)
@@ -8,9 +7,9 @@ export default function AdoptionList() {
     <>
       <div className="">
         <h1>Essa é a pagina de controle de adocao</h1>
-        <Button onClick={() => setCount((count) => count + 1)}>
+        <button onClick={() => setCount((count) => count + 1)}>
           count is {count}
-        </Button>
+        </button>
       </div>
 
     </>
